shipcat reconcile
This is a huge restructure to allow upgrade to run in parallel.

It pulls in the threadpool dependency to run helm::upgrade for each
pre-generated helm values files (it will generate them all first
sequentially), then it will run upgrade on them.

This is currently in dry-run mode only as it is so new.

Have tested the normal helm stuff in shipcat and it all works. Types
complained a lot, but they also fixed it for me.

Most annoying thing was having to pull out the boxed templating function
and instead pass references to the templating engine.