Common Container struct (#168)
Extract a `Container` struct (and `ContainerSource`) which represents the common properties of `Sidecar`, `InitContainer`, `Worker`, `Job` and `CronJob`. This also standardises the properties of them all (e.g., they all have `env` and `preserveEnv` properties, fixing #167).

I've also removed the `preserveEnv` property from `shipcat_definitions`, the behaviour of this is now handled in `shipcat_filebacked`, by merging the Worker's (etc) env with the parent when building the main structs (so it no longer requires custom logic in the Helm templates).

Also introduce a `Build` trait for the `*Source` structs, to make it simpler to build sources into the main structs (it may be possible to derive this in the future).

There shouldn't be any breaking changes in this (`preserveEnv` is no longer included in the Helm values, but that's equivalent to `preserveEnv: false`).