Adding pod-annotations to shipcat (#194)
Need the ability to add pod annotations to deployments, cronJobs, etc. to enable tools like kube2iam.

Not sure if this kind of functionality should have tests added to it, since the most logic is on the helm templating side of things. Tried to copy the pattern for `labels` and `service_annotations`. 

Let me know if any other changes are needed