[DXKP-120] Add kompass-hub plugin (#475)
* [DXKP-119] Add new endpoint for kompass-hub

This patch adds a new endpoint to raftcat as a preparation step for
running it as a kompass plugin.

The endpoint translates the current manifests to the format expected by
kompass-hub. The support for registering raftcat with the hub so that
this endpoint is used will be added in a different PR.

* [DXKP-120] Register with kompass-hub

This patch registers raftcat with kompass-hub and sends a heartbeat
every 30 seconds.

* [DXKP-119] Parse URL